# Conflicts:
#	introspection-engine/connectors/sql-introspection-connector/src/calculate_datamodel.rs
#	introspection-engine/connectors/sql-introspection-connector/tests/unit_tests/postgres.rs

+++ migration-engine/connectors/sql-migration-connector/src/sql_destructive_changes_checker.rs
@@ -4,12 +4,14 @@ use crate::{
 };
 use migration_connector::*;
 use quaint::{ast::*, prelude::Queryable};
+use sql_schema_describer::SqlSchemaDescriberBackend;
 use std::sync::Arc;
 
 pub struct SqlDestructiveChangesChecker {
     pub connection_info: ConnectionInfo,
     pub schema_name: String,
     pub database: Arc<dyn Queryable + Send + Sync>,
+    pub database_describer: Arc<dyn SqlSchemaDescriberBackend + Send + Sync + 'static>,
 }
 
 impl SqlDestructiveChangesChecker {
@@ -93,19 +95,32 @@ impl SqlDestructiveChangesChecker {
     async fn check_alter_column(
         &self,
         alter_column: &AlterColumn,
-        table: &sql_schema_describer::Table,
+        previous_table: &sql_schema_describer::Table,
         diagnostics: &mut DestructiveChangeDiagnostics,
     ) -> SqlResult<()> {
-        let values_count = self.count_values_in_column(&alter_column.name, table).await?;
+        let values_count = self.count_values_in_column(&alter_column.name, previous_table).await?;
+        let previous_column = previous_table
+            .column(&alter_column.name)
+            .expect("unsupported column renaming");
 
         if values_count > 0 {
             diagnostics.add_warning(MigrationWarning {
                 description: format!(
                                  "You are about to alter the column `{column_name}` on the `{table_name}` table, which still contains {values_count} non-null values. The data in that column will be lost.",
                                  column_name=alter_column.name,
-                                 table_name=&table.name,
+                                 table_name=&previous_table.name,
                                  values_count=values_count,
                              )
+            })
+        } else if previous_table.is_part_of_foreign_key(&alter_column.column.name)
+            && alter_column.column.default.is_none()
+            && previous_column.default.is_some()
+        {
+            diagnostics.add_warning(MigrationWarning {
+                description: format!(
+                    "The migration is about to remove a default value on the foreign key field `{}.{}`.",
+                    previous_table.name, alter_column.name,
+                ),
             })
         }
 
@@ -120,27 +135,21 @@ impl SqlDestructiveChangesChecker {
                 SqlMigrationStep::AlterTable(alter_table) => {
                     // The table in alter_table is the updated table, but we want to
                     // check against the current state of the table.
-                    let before_table = database_migration
-                        .before
-                        .get_table(&alter_table.table.name)
-                        .ok_or_else(|| {
-                            SqlError::Generic(format!(
-                                "Internal Error: altering previously-unknown table {}",
-                                &alter_table.table.name
-                            ))
-                        })?;
+                    let before_table = database_migration.before.get_table(&alter_table.table.name);
 
-                    for change in &alter_table.changes {
-                        match *change {
-                            TableChange::DropColumn(ref drop_column) => {
-                                self.check_column_drop(drop_column, before_table, &mut diagnostics)
-                                    .await?
+                    if let Some(before_table) = before_table {
+                        for change in &alter_table.changes {
+                            match *change {
+                                TableChange::DropColumn(ref drop_column) => {
+                                    self.check_column_drop(drop_column, before_table, &mut diagnostics)
+                                        .await?
+                                }
+                                TableChange::AlterColumn(ref alter_column) => {
+                                    self.check_alter_column(alter_column, before_table, &mut diagnostics)
+                                        .await?
+                                }
+                                _ => (),
                             }
-                            TableChange::AlterColumn(ref alter_column) => {
-                                self.check_alter_column(alter_column, before_table, &mut diagnostics)
-                                    .await?
-                            }
-                            _ => (),
                         }
                     }
                 }
